feat(*): Another new (old) direction, this time with some help incoming.
Added Kevin to "author", brought back a lot of old code, and compendiums, but this time in the ce directory (as these are intended for ce), we need to decide how to handle other variants (maybe simply name them like 'ce-skills'? Or, maybe foundry supports folders for compendiums, I *think* I've seen that in a system.

Figured this was enough for a bump of minor version.