feat: add damage rolls from ship sheet and add prompt for all direct damage rolls (#899)
This release adds the ability to roll from the damage field of a component on the ship sheet.  It must be a valid roll formula. 

All direct damage rolls (both traveller and ship items) now accommodate the shift-click setting.  By default, a popup window opens to the user to adjust the roll formula. Shift-click performs an immediate roll.  This behavior is reversed if the setting is true.