feat: Custom status effects icons (#1403)
This release adds custom status icons for Cepheus/Traveller.  A display setting has been added to hide status icons that are for more fantasy settings.

Also, this release fixes a bug for players not being able to drag-drop consumables on items. Thanks to Bryshen for identifying this issue.

And some localization updates courtesy of @ForjaSalvaje